Match 5 functions in Menu/

+++ src/Menu/sub_80046B74.c
@@ -2,7 +2,7 @@
 
 extern RECT rect_800AB6C8;
 
-void sub_80046B74(void *a1, short a2)
+void sub_80046B74(void *a1, int a2)
 {
     RECT v3;
